+++ packages/connector/src/zkp/consts/index.ts
@@ -1,2 +1,1 @@
-export * from './snap';
 export * from './general';