+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 [Unreleased]
 
+## [2.1.1] - 2024-06-18
+
+### Fixed
+- `@rarimo/rarime-connector` - request params
+
+[Unreleased]
+
 ## [2.1.0] - 2024-04-08
 
 ### Added
@@ -194,7 +201,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Implemented `@rarimo/rarime-connector` and `@rarimo/rarime` packages
 
-[Unreleased]: https://github.com/rarimo/rarime/compare/2.1.0...HEAD
+[Unreleased]: https://github.com/rarimo/rarime/compare/2.1.1...HEAD
+[2.1.1]: https://github.com/rarimo/rarime/compare/2.1.0...2.1.1
 [2.1.0]: https://github.com/rarimo/rarime/compare/2.0.3...2.1.0
 [2.0.3]: https://github.com/rarimo/rarime/compare/2.0.2...2.0.3
 [2.0.2]: https://github.com/rarimo/rarime/compare/2.0.1...2.0.2